Info Overlay added.

+++ app/src/main/java/com/mylozo/presenti/app/view/ScanQRActivity.kt
@@ -10,13 +10,11 @@ import android.os.Bundle
 import android.text.TextUtils
 import android.util.Log
 import android.view.View
-import android.widget.Button
-import android.widget.ImageView
-import android.widget.ProgressBar
-import android.widget.TextView
+import android.widget.*
 import androidx.activity.result.contract.ActivityResultContracts.StartActivityForResult
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.app.ActivityCompat
 import com.google.android.material.snackbar.Snackbar
 import com.mylozo.presenti.app.R
@@ -25,11 +23,12 @@ import com.mylozo.presenti.app.presenter.NetworkResponseListener
 import java.io.IOException
 
 
-class ScanQRActivity : AppCompatActivity(), NetworkResponseListener {
+class ScanQRActivity : AppCompatActivity(), NetworkResponseListener, View.OnClickListener {
 
     val REQUEST_CAMERA_PERMISSION = 700
     val REQUEST_LOCATION_PERMISSION = 701
     var isFromLocation = false
+    var overlayClicks = 0;
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -215,9 +214,15 @@ class ScanQRActivity : AppCompatActivity(), NetworkResponseListener {
 
     private fun initUI() {
 
-        if (EmployeePrefs.getValid(this)) {
-            EmployeeRepository.employee = EmployeePrefs.getEmployeeDetails(this)
-            EmployeeRepository.business = EmployeePrefs.getBusinessDetails(this)
+        if(EmployeePrefs.getInfoOverlay(this))
+            findViewById<RelativeLayout>(R.id.intro).visibility = View.GONE
+
+        try {
+            if (EmployeePrefs.getValid(this)) {
+                EmployeeRepository.employee = EmployeePrefs.getEmployeeDetails(this)
+                EmployeeRepository.business = EmployeePrefs.getBusinessDetails(this)
+            }
+        } catch (e: Exception) {
         }
 
         findViewById<TextView>(R.id.company_name).text =
@@ -225,28 +230,43 @@ class ScanQRActivity : AppCompatActivity(), NetworkResponseListener {
 
         val qrImage = findViewById<ImageView>(R.id.barcode)
         qrImage.setOnClickListener {
-            validatePermissions()
+            if (EmployeePrefs.getInfoOverlay(this))
+                validatePermissions()
 
         }
 
         val buttonLogOut = findViewById<ImageView>(R.id.log_out)
         buttonLogOut.setOnClickListener {
-            EmployeePrefs.deleteAllDetails(this@ScanQRActivity)
-            finish()
+            if (EmployeePrefs.getInfoOverlay(this)) {
+                EmployeePrefs.deleteAllDetails(this@ScanQRActivity)
+                finish()
+            }
         }
 
         val buttonRemoteLogin = findViewById<Button>(R.id.remote)
         buttonRemoteLogin.isEnabled =
             EmployeeRepository.business.data?.isBusinessLocationDetect != 1
         buttonRemoteLogin.setOnClickListener {
-            checkLocationPermission(true)
+            if (EmployeePrefs.getInfoOverlay(this))
+                checkLocationPermission(true)
         }
 
         findViewById<TextView>(R.id.user_name).text =
             "Hi, " + EmployeeRepository.employee.data?.empName
 
         findViewById<ImageView>(R.id.language).setOnClickListener {
-            localeActivityResultLauncher.launch(Intent(this, LanguageSelectorActivity::class.java))
+            if (EmployeePrefs.getInfoOverlay(this))
+                localeActivityResultLauncher.launch(
+                    Intent(
+                        this,
+                        LanguageSelectorActivity::class.java
+                    )
+                )
+        }
+
+        if (!EmployeePrefs.getInfoOverlay(this)) {
+            findViewById<RelativeLayout>(R.id.intro).bringToFront()
+            findViewById<TextView>(R.id.next_overlay).setOnClickListener(this)
         }
     }
 
@@ -295,6 +315,43 @@ class ScanQRActivity : AppCompatActivity(), NetworkResponseListener {
             }
         } catch (e: IOException) {
             e.printStackTrace()
+        }
+    }
+
+    override fun onClick(v: View?) {
+        if (v?.id == R.id.next_overlay && !EmployeePrefs.getInfoOverlay(this)) {
+            if (overlayClicks == 0) {
+                overlayClicks++
+
+                findViewById<ImageView>(R.id.overlay_lang).visibility = View.INVISIBLE
+                findViewById<TextView>(R.id.txt_language).visibility = View.INVISIBLE
+
+                findViewById<ImageView>(R.id.overlay_logout).visibility = View.VISIBLE
+                findViewById<TextView>(R.id.click_logout).visibility = View.VISIBLE
+            } else if (overlayClicks == 1) {
+                overlayClicks++
+                findViewById<ImageView>(R.id.overlay_qr).visibility = View.VISIBLE
+                findViewById<TextView>(R.id.txt_scan_qr).visibility = View.VISIBLE
+
+                findViewById<ImageView>(R.id.overlay_logout).visibility = View.INVISIBLE
+                findViewById<TextView>(R.id.click_logout).visibility = View.INVISIBLE
+            } else if (overlayClicks == 2) {
+                overlayClicks++
+                findViewById<ImageView>(R.id.overlay_logremote).visibility = View.VISIBLE
+                findViewById<TextView>(R.id.txt_logremote).visibility = View.VISIBLE
+
+                findViewById<ImageView>(R.id.overlay_qr).visibility = View.INVISIBLE
+                findViewById<TextView>(R.id.txt_scan_qr).visibility = View.INVISIBLE
+            } else if (overlayClicks == 3) {
+
+                findViewById<ImageView>(R.id.overlay_logremote).visibility = View.INVISIBLE
+                findViewById<TextView>(R.id.txt_logremote).visibility = View.INVISIBLE
+
+                findViewById<RelativeLayout>(R.id.intro).visibility = View.GONE
+                findViewById<ConstraintLayout>(R.id.rootView).bringToFront()
+                EmployeePrefs.setInfoOverlay(this,true)
+            }
+
         }
     }
 }
